Island: Move Login endpoint to authentication_service
