+++ monkey/monkey_island/cc/resources/auth/__init__.py
@@ -1,3 +1,2 @@
-from .login import Login
 from .logout import Logout
 from .registration_status import RegistrationStatus
