+++ monkey/tests/unit_tests/monkey_island/conftest.py
@@ -14,7 +14,8 @@ import monkey_island
 from common import AccountRole
 from common.utils.code_utils import insecure_generate_random_string
 from monkey_island.cc.flask_utils import AbstractResource
-from monkey_island.cc.resources.auth import Login, Logout
+from monkey_island.cc.resources.auth import Logout
+from monkey_island.cc.services.authentication_service.flask_resources.login import Login
 from monkey_island.cc.services.authentication_service.flask_resources.register import Register
 from monkey_island.cc.services.authentication_service.role import Role
 from monkey_island.cc.services.authentication_service.user import User
